rename author/book package names
