+++ molgenis-test/src/main/java/org/molgenis/test/data/OneToManyTestHarness.java
@@ -11,22 +11,22 @@ import org.molgenis.test.data.staticentity.bidirectional.person3.PersonFactory3;
 import org.molgenis.test.data.staticentity.bidirectional.person3.PersonMetaData3;
 import org.molgenis.test.data.staticentity.bidirectional.person4.PersonFactory4;
 import org.molgenis.test.data.staticentity.bidirectional.person4.PersonMetaData4;
-import org.molgenis.test.data.staticentity.bidirectional.test1.AuthorFactory1;
-import org.molgenis.test.data.staticentity.bidirectional.test1.AuthorMetaData1;
-import org.molgenis.test.data.staticentity.bidirectional.test1.BookFactory1;
-import org.molgenis.test.data.staticentity.bidirectional.test1.BookMetaData1;
-import org.molgenis.test.data.staticentity.bidirectional.test2.AuthorFactory2;
-import org.molgenis.test.data.staticentity.bidirectional.test2.AuthorMetaData2;
-import org.molgenis.test.data.staticentity.bidirectional.test2.BookFactory2;
-import org.molgenis.test.data.staticentity.bidirectional.test2.BookMetaData2;
-import org.molgenis.test.data.staticentity.bidirectional.test3.AuthorFactory3;
-import org.molgenis.test.data.staticentity.bidirectional.test3.AuthorMetaData3;
-import org.molgenis.test.data.staticentity.bidirectional.test3.BookFactory3;
-import org.molgenis.test.data.staticentity.bidirectional.test3.BookMetaData3;
-import org.molgenis.test.data.staticentity.bidirectional.test4.AuthorFactory4;
-import org.molgenis.test.data.staticentity.bidirectional.test4.AuthorMetaData4;
-import org.molgenis.test.data.staticentity.bidirectional.test4.BookFactory4;
-import org.molgenis.test.data.staticentity.bidirectional.test4.BookMetaData4;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook1.AuthorFactory1;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook1.AuthorMetaData1;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook1.BookFactory1;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook1.BookMetaData1;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook2.AuthorFactory2;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook2.AuthorMetaData2;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook2.BookFactory2;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook2.BookMetaData2;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook3.AuthorFactory3;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook3.AuthorMetaData3;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook3.BookFactory3;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook3.BookMetaData3;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook4.AuthorFactory4;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook4.AuthorMetaData4;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook4.BookFactory4;
+import org.molgenis.test.data.staticentity.bidirectional.authorbook4.BookMetaData4;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
