+++ molgenis-test/src/main/java/org/molgenis/test/data/staticentity/bidirectional/authorbook1/BookMetaData1.java
@@ -1,4 +1,4 @@
-package org.molgenis.test.data.staticentity.bidirectional.test4;
+package org.molgenis.test.data.staticentity.bidirectional.authorbook1;
 
 import org.molgenis.data.meta.SystemEntityMetaData;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,21 +12,21 @@ import static org.molgenis.data.meta.model.Package.PACKAGE_SEPARATOR;
 import static org.molgenis.data.system.model.RootSystemPackage.PACKAGE_SYSTEM;
 
 /**
- * AuthorMetaData4 and BookMetaData4 define two entities with a OneToMany relation of which both sides are required.
+ * AuthorMetaData1 and BookMetaData1 define two entities with a OneToMany relation of which no side is required.
  */
 @Component
-public class BookMetaData4 extends SystemEntityMetaData
+public class BookMetaData1 extends SystemEntityMetaData
 {
-	private static final String SIMPLE_NAME = "Book4";
+	private static final String SIMPLE_NAME = "Book1";
 	public static final String MY_REF_ENTITY = PACKAGE_SYSTEM + PACKAGE_SEPARATOR + SIMPLE_NAME;
 
 	public static final String ID = "id";
 	public static final String LABEL = "label";
 	public static final String AUTHOR = "author";
 
-	private AuthorMetaData4 authorMetaData;
+	private AuthorMetaData1 authorMetaData;
 
-	BookMetaData4()
+	BookMetaData1()
 	{
 		super(SIMPLE_NAME, PACKAGE_SYSTEM);
 	}
@@ -36,13 +36,13 @@ public class BookMetaData4 extends SystemEntityMetaData
 	{
 		setLabel("Book");
 
-		addAttribute(ID, ROLE_ID).setAuto(true).setLabel("Identifier");
+		addAttribute(ID, ROLE_ID).setLabel("Identifier");
 		addAttribute(LABEL, ROLE_LABEL).setNillable(true).setLabel("Label");
-		addAttribute(AUTHOR).setDataType(XREF).setRefEntity(authorMetaData).setNillable(false);
+		addAttribute(AUTHOR).setDataType(XREF).setRefEntity(authorMetaData);
 	}
 
 	@Autowired
-	public void setAuthorMetaData(AuthorMetaData4 authorMetaData)
+	public void setAuthorMetaData(AuthorMetaData1 authorMetaData)
 	{
 		this.authorMetaData = requireNonNull(authorMetaData);
 	}
