+++ molgenis-test/src/main/java/org/molgenis/test/data/staticentity/bidirectional/authorbook4/AuthorFactory4.java
@@ -1,4 +1,4 @@
-package org.molgenis.test.data.staticentity.bidirectional.test4;
+package org.molgenis.test.data.staticentity.bidirectional.authorbook4;
 
 import org.molgenis.data.AbstractSystemEntityFactory;
 import org.molgenis.test.data.staticentity.bidirectional.Author;
